[fix bug 1029829][fix bug 1029838][bug 1081917] Thunderbird RNA
* Created releasenotes bedrock app
* Migrated firefox release notes views to releasenotes app.
* Added thunderbird releasenotes urls
* Added global.conf RewriteRules for Thunderbird release notes.
* Thunderbird base, release notes, system requirements templates
