+++ bedrock/thunderbird/templates/thunderbird/base-resp.html
@@ -1,0 +1,86 @@
+{# This Source Code Form is subject to the terms of the Mozilla Public
+ # License, v. 2.0. If a copy of the MPL was not distributed with this
+ # file, You can obtain one at http://mozilla.org/MPL/2.0/. -#}
+
+{% extends "base-resp.html" %}
+
+{% block body_class %}sky{% endblock %}
+
+{% block site_css %}
+  {{ css('firefox-resp') }}
+{% endblock %}
+
+{% block site_header_nav %}
+<nav id="nav-main" role="navigation">
+  <span class="toggle" role="button" aria-controls="nav-main-menu" aria-expanded="false" tabindex="0">{{_('Menu')}}</span>
+  <ul id="nav-main-menu" class="has-submenus">
+    <li class="first" id="nav-main-features"><a href="/thunderbird/" tabindex="0" aria-owns="nav-main-features-submenu" aria-haspopup="true">{{_('Thunderbird')}}</a>
+      <ul aria-expanded="false" id="nav-main-features-submenu" class="submenu">
+        <li class="first">
+          <a href="/thunderbird/channel/" tabindex="-1">
+            {{_('Releases')}}
+          </a>
+        </li>
+        <li>
+          <a href="/thunderbird/features/" tabindex="-1">
+            {{_('Features')}}
+          </a>
+        </li>
+        <li class="last">
+          <a href="/thunderbird/organizations/" tabindex="-1">
+            {{_('Thunderbird for Organizations')}}
+          </a>
+        </li>
+      </ul>
+    </li>
+    <li id="nav-main-addons">
+      <a href="https://addons.mozilla.org/" tabindex="0" aria-owns="nav-main-addons-submenu" aria-haspopup="true">
+        {{_('Add-ons')}}
+      </a>
+      <ul aria-expanded="false" id="nav-main-addons-submenu" class="submenu">
+        <li class="first">
+          <a href="https://addons.mozilla.org/thunderbird/extensions/" tabindex="-1">
+            {{_('Add-ons')}}
+          </a>
+        </li>
+        <li class="last">
+          <a href="https://addons.mozilla.org/thunderbird/themes/" tabindex="-1">
+            {{_('Themes')}}
+          </a>
+        </li>
+      </ul>
+    </li>
+    <li class="last" id="nav-main-support">
+      <a href="https://support.mozilla.org/products/thunderbird/" tabindex="0"
+         aria-owns="nav-main-addons-submenu" aria-haspopup="true">
+        {{_('Support')}}
+      </a>
+      <ul aria-expanded="false" id="nav-main-addons-submenu" class="submenu">
+        <li class="first last">
+          <a href="https://support.mozilla.org/products/thunderbird" tabindex="-1">
+            {{_('Thunderbird Support')}}
+          </a>
+        </li>
+      </ul>
+    </li>
+  </ul>
+</nav>
+{% endblock %}
+
+{% block email_form %}
+{% endblock email_form %}
+
+{% block site_header_logo %}
+<h2>
+  <a href="/thunderbird/">
+    <img src="{{ media('img/thunderbird/template/thunderbird.png?2014-10') }}" 
+         alt="Mozilla Thunderbird" 
+         height="70" 
+         width="260">
+  </a>
+</h2>
+{% endblock %}
+
+{% block site_js %}
+  {{ js('firefox-resp') }}
+{% endblock %}
