+++ bedrock/thunderbird/templates/thunderbird/releases/beta-notes.html
@@ -1,0 +1,7 @@
+{# This Source Code Form is subject to the terms of the Mozilla Public
+ # License, v. 2.0. If a copy of the MPL was not distributed with this
+ # file, You can obtain one at http://mozilla.org/MPL/2.0/. #}
+
+{% extends "thunderbird/releases/release-notes.html" %}
+
+{% set channel_name = 'Beta' %}
