+++ bedrock/thunderbird/templates/thunderbird/releases/index.html
@@ -1,0 +1,72 @@
+{# This Source Code Form is subject to the terms of the Mozilla Public
+ # License, v. 2.0. If a copy of the MPL was not distributed with this
+ # file, You can obtain one at http://mozilla.org/MPL/2.0/. #}
+
+{% extends "thunderbird/base-resp.html" %}
+
+{% block page_title %}{{ _('Mozilla Thunderbird Release Notes') }}{% endblock %}
+{% block page_desc %}{{ _('Release notes for each version of Thunderbird.') }}{% endblock %}
+{% block body_id %}thunderbird-releases-index{% endblock %}
+
+{% block site_css %}
+  {{ css('firefox_releases_index') }}
+{% endblock %}
+
+{% block breadcrumbs %}
+  <nav class="breadcrumbs">
+    <a href="{{ url('mozorg.home') }}">{{ _('Home') }}</a> >
+    <a href="/thunderbird/">{{ _('Thunderbird') }}</a> >
+    <span>{{ _('Release Notes') }}</span>
+  </nav>
+{% endblock %}
+
+{% block content %}
+  <div id="main-feature">
+    <h1>{{ _('Thunderbird Releases') }}</h1>
+    <p>{{ _('Thunderbird release notes are specific to each version of the application. Select your version from the list below to see the release notes for it.') }}</p>
+  </div>
+
+  <div id="main-content">
+    <ol reversed>
+      {% for int_version, versions in releases -%}
+      <li>
+        <strong>{{ get_link(int_version, versions.major) }}</strong>
+        {% if versions.minor -%}
+        <ol>
+          {% for version in versions.minor -%}
+          <li>{{ get_link(int_version, version) }}</li>
+          {% endfor -%}
+        </ol>
+        {% endif -%}
+      </li>
+      {% endfor -%}
+      <li>
+        <strong><a href="0.1.html">0.1</a></strong>
+        <ol>
+          <li><a href="0.2.html">0.2</a></li>
+          <li><a href="0.3.html">0.3</a></li>
+          <li><a href="0.4.html">0.4</a></li>
+          <li><a href="0.5.html">0.5</a></li>
+          <li><a href="0.6.html">0.6</a></li>
+          <li><a href="0.7.html">0.7</a></li>
+          <li><a href="0.8.html">0.8</a></li>
+          <li><a href="0.9.html">0.9</a></li>
+        </ol>
+      </li>
+    </ol>
+  </div>
+{% endblock %}
+
+{% macro get_link(int_version, version) %}
+  {%- if int_version < 2 -%}
+    <a href="{{ version }}.html">{{ version }}</a>
+  {%- elif version == '2.0' -%}
+    <a href="../2.0.0.0/releasenotes/">{{ version }}</a>
+  {%- elif version == '17.0.9' -%}
+    <a href="../17.0.9esr/releasenotes/">{{ version }}</a>
+  {%- elif version == '17.0.10' -%}
+    <a href="../17.0.10esr/releasenotes/">{{ version }}</a>
+  {%- else -%}
+    <a href="../{{ version }}/releasenotes/">{{ version }}</a>
+  {%- endif -%}
+{% endmacro %}
