+++ bedrock/thunderbird/templates/thunderbird/releases/release-notes.html
@@ -1,0 +1,161 @@
+{# This Source Code Form is subject to the terms of the Mozilla Public
+ # License, v. 2.0. If a copy of the MPL was not distributed with this
+ # file, You can obtain one at http://mozilla.org/MPL/2.0/. #}
+
+{% extends "thunderbird/base-resp.html" %}
+
+{% block page_title_prefix %}{% endblock %}
+{% block page_title %}{{ _('Thunderbird — {channel} Notes ({version})')|f(channel=channel_name, version=release.version) }}{% endblock %}
+
+{% block body_id %}notes{% endblock %}
+{% block body_class %}thunderbird-notes sky{% endblock %}
+
+{% block site_css %}
+  {{ css('firefox_releasenotes') }}
+{% endblock %}
+
+{# channel_name is for display purposes where needed.  #}
+{% set channel_name = channel_name|default('') %}
+
+{% block content %}
+<main role="main">
+  <article id="main-content">
+
+  {% block notes_header %}
+    <header class="notes-head">
+    <h1>{{ _('Thunderbird {channel} Notes')|f(channel=channel_name) }}</h1>
+    <h2>{{ _('Version {version}, first offered to {channel} channel users on {date}')|f(channel=release.channel, date=release.release_date|l10n_format_date, version=release.version) }}</h2>
+
+      <p>
+        {{ _('Check out "What’s New" and "Known Issues" for this version of Thunderbird below.') }}
+      {% trans feedback='https://support.mozilla.org/questions/new/thunderbird',
+               bugzilla='https://bugzilla.mozilla.org/' %}
+        As always, you’re encouraged to
+        <a href="{{ feedback }}">tell us what you think</a>,
+        or <a href="{{ bugzilla }}">file a bug in Bugzilla</a>.
+      {% endtrans %}
+      {% if release.bug_list %}
+      <form action="https://bugzilla.mozilla.org/buglist.cgi" method="post">
+        <input type="hidden" name="limit" value="0">
+        <input type="hidden" name="bug_id" value="{{ release.bug_list }}">
+        <p>If interested, please see the <button type="submit">complete list of changes</button> in this release.</p>
+      </form>
+      {% else %}
+          If interested, please see the  <a href="{{ release.get_bug_search_url() }}">complete list of changes</a> in this release.
+      {% endif %}
+        {{ release.text|markdown|safe }}
+      </p>
+    </header>
+  {% endblock %}
+
+    <div class="main-column">
+      {% if new_features %}
+      <section class="notes-section" id="new-features">
+        <h3>{{ _('What’s New') }}</h3>
+        <ul class="section-items tagged">
+          {% for note in new_features %}
+          <li {% if not note.tag %}class="untagged"{% endif %} id="note-{{ note.id }}">
+            {% if note.tag %}
+              <b class="tag tag-{{ note.tag.lower() }}">{{ note.tag }}</b>
+            {% endif %}
+              {{ note.note|markdown|safe }}
+          </li>
+          {% endfor %}
+        </ul>
+      </section>
+      {% endif %}
+      {% if known_issues %}
+      <section class="notes-section" id="known-issues">
+        <h3>{{ _('Known Issues') }}</h3>
+        <ul class="section-items tagged">
+          {% for note in known_issues %}
+            <li id="note-{{ note.id }}">
+              <b class="tag tag-unresolved">{{ _('unresolved') }}</b>
+              {{ note.note|markdown|safe }}
+            {% if note.fixed_in_release %}
+              <p class="note">
+                <a href="{{ releasenotes_url(note.fixed_in_release) }}">
+                  {{ _('Resolved in v{version_number}')|f(version_number=note.fixed_in_release.version) }}
+                </a>
+              </p>
+            {% endif %}
+            </li>
+          {% endfor %}
+        </ul>
+      </section>
+      {% endif %}
+    </div>
+
+  {% block notes_sidebar %}
+    <aside id="sidebar" class="sidebar">
+      <section id="try">
+        <h3>{{ _('Try Thunderbird %s')|format(channel_name) }}</h3>
+        <ol>
+          <li class="try1">
+            <a href="{{ url('thunderbird.system_requirements', version) }}">
+              {{ _('Check') }}
+            </a>
+          </li>
+          <li class="try2">
+            <a href="{{ download_url }}">
+              {{ _('Download') }}
+            </a>
+          </li>
+          <li class="try3">
+            <a href="https://support.mozilla.org/kb/Installing%20Thunderbird">
+              {{ _('Install') }}
+            </a>
+          </li>
+          <li class="try4">
+            <a href="https://addons.mozilla.org/thunderbird/">
+              {{ _('Customize') }}
+            </a>
+          </li>
+        </ol>
+      </section>
+
+      <section id="get-involved">
+        <h3>{{ _('Want to get involved?') }}</h3>
+        <p>{{ _('Did you know that most of the content in Thunderbird Support was written by volunteers?') }}<br>
+        <a href="{{ url('mozorg.contribute') }}">{{ _('Find out more') }}</a></p>
+      </section>
+
+      <section id="problems">
+        <h3>{{ _('Having Problems?') }}</h3>
+        <ol>
+          <li>
+            <a href="{{ support_url }}">
+              {{ _('Search for answers on the Thunderbird Support page') }}
+            </a>
+          </li>
+          {% if known_issues %}
+            <li>
+              <a href="#known-issues">
+                {{ _('Look at the known issues list and see if we already know about the problem') }}
+              </a>
+            </li>
+          {% endif %}
+          <li>
+            <a href="https://bugzilla.mozilla.org/enter_bug.cgi">
+              {{ _('If you can’t find reference to your issue, please consider filing a technical bug report') }}
+            </a>
+          </li>
+        </ol>
+      </section>
+
+      <section id="other">
+        <h3>{{ _('Other Resources') }}</h3>
+        <p>
+          <a href="https://blog.mozilla.com/thunderbird/">
+            {{ _('The Thunderbird Blog') }}
+          </a>
+        </p>
+      </section>
+    </aside>
+  {% endblock %}
+
+  </article>
+</main>
+{% endblock %}
+
+{% block js %}{% endblock  %}
