+++ etc/httpd/global.conf
@@ -566,6 +566,16 @@ RewriteRule ^/(\w{2,3}(?:-\w{2}(?:-mac)?)?/)?(firefox|mobile)/(2[38]\.0(?:\.\d)?
 RewriteRule ^/(\w{2,3}(?:-\w{2}(?:-mac)?)?/)?(firefox|mobile)/(29\.0(?:beta|\.\d)?)/releasenotes(/)?$ /b/$1$2/$3/releasenotes$4 [L,PT]
 RewriteRule ^/(\w{2,3}(?:-\w{2}(?:-mac)?)?/)?(firefox|mobile)/([3-9]\d\.\d(?:a2|beta|\.\d)?)/(aurora|release)notes(/)?$ /b/$1$2/$3/$4notes$5 [L,PT]
 
+
+
+# bug 1081917, 1029829, 1029838
+RewriteRule ^/(\w{2,3}(?:-\w{2}(?:-mac)?)?/)?thunderbird/([3-9]\d\.\d(?:a2|beta|\.\d)?)/releasenotes(/)?$ /b/$1thunderbird/$2/releasenotes$3 [L,PT]
+RewriteRule ^/(\w{2,3}(?:-\w{2}(?:-mac)?)?/)?thunderbird/([3-9]\d\.\d(?:a2|beta|\.\d)?)/system-requirements(/)?$ /b/$1thunderbird/$2/system-requirements$3 [L,PT]
+RewriteRule ^/thunderbird/releases/(.*)$ http://www-archive.mozilla.org/thunderbird_releasenotes/en-US/thunderbird/releases/$1 [L,R=301]
+RewriteRule ^/(\w{2,3}(?:-\w{2}(?:-mac)?)?/)thunderbird/(.*)/releasenotes(.*)$ http://www-archive.mozilla.org/thunderbird_releasenotes/en-US/thunderbird/$2/releasenotes$4 [L,R=301]
+RewriteRule ^/(\w{2,3}(?:-\w{2}(?:-mac)?)?/)thunderbird/(.*)/system-requirements(.*)$ http://www-archive.mozilla.org/thunderbird_releasenotes/en-US/thunderbird/$2/system-requirements$4 [L,R=301]
+
+
 # bug 1001238, 1025056
 RewriteRule ^/(\w{2,3}(?:-\w{2}(?:-mac)?)?/)?firefox/(24\.[5678]\.\d)/releasenotes(/)?$ /b/$1firefox/$2/releasenotes$3 [L,PT]
